Make sure we're using float fields

+++ BrianJacobs/Scheduler/database/migrations/2014_07_12_180544_create_user_credits.php
@@ -17,8 +17,8 @@ class CreateUserCredits extends Migration {
 			$table->bigIncrements('id');
 			$table->string('code', 12);
 			$table->string('type');
-			$table->integer('value')->default(0);
-			$table->integer('claimed')->default(0);
+			$table->float('value')->default(0);
+			$table->float('claimed')->default(0);
 			$table->integer('user_id')->unsigned()->nullable();
 			$table->string('email')->nullable();
 			$table->text('notes')->nullable();
